feat: update client packet decoders
