+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/buttons/IfButtonDDecoder.kt
@@ -16,12 +16,12 @@ public class IfButtonDDecoder : ProxyMessageDecoder<IfButtonD> {
         buffer: JagByteBuf,
         session: Session,
     ): IfButtonD {
-        val selectedSub = buffer.g2Alt1()
-        val targetCombinedId = buffer.gCombinedIdAlt1()
-        val selectedCombinedId = buffer.gCombinedIdAlt3()
         val selectedObj = buffer.g2Alt1()
-        val targetObj = buffer.g2()
+        val targetObj = buffer.g2Alt1()
+        val selectedCombinedId = buffer.gCombinedIdAlt1()
         val targetSub = buffer.g2Alt2()
+        val selectedSub = buffer.g2Alt3()
+        val targetCombinedId = buffer.gCombinedIdAlt3()
         return IfButtonD(
             selectedCombinedId,
             selectedSub,
