+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/buttons/IfButtonTDecoder.kt
@@ -2,7 +2,7 @@ package net.rsprox.protocol.v230.game.incoming.decoder.codec.buttons
 
 import net.rsprot.buffer.JagByteBuf
 import net.rsprot.protocol.ClientProt
-import net.rsprot.protocol.util.gCombinedIdAlt1
+import net.rsprot.protocol.util.gCombinedIdAlt2
 import net.rsprot.protocol.util.gCombinedIdAlt3
 import net.rsprox.protocol.ProxyMessageDecoder
 import net.rsprox.protocol.game.incoming.model.buttons.IfButtonT
@@ -16,12 +16,12 @@ public class IfButtonTDecoder : ProxyMessageDecoder<IfButtonT> {
         buffer: JagByteBuf,
         session: Session,
     ): IfButtonT {
-        val targetObj = buffer.g2Alt3()
-        val selectedSub = buffer.g2Alt2()
-        val targetSub = buffer.g2Alt1()
-        val selectedObj = buffer.g2()
-        val targetCombinedId = buffer.gCombinedIdAlt1()
-        val selectedCombinedId = buffer.gCombinedIdAlt3()
+        val selectedSub = buffer.g2Alt3()
+        val targetCombinedId = buffer.gCombinedIdAlt3()
+        val targetObj = buffer.g2Alt1()
+        val selectedCombinedId = buffer.gCombinedIdAlt2()
+        val targetSub = buffer.g2()
+        val selectedObj = buffer.g2Alt3()
         return IfButtonT(
             selectedCombinedId,
             selectedSub,
