+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/events/EventCameraPositionDecoder.kt
@@ -14,8 +14,8 @@ public class EventCameraPositionDecoder : ProxyMessageDecoder<EventCameraPositio
         buffer: JagByteBuf,
         session: Session,
     ): EventCameraPosition {
-        val angleY = buffer.g2()
-        val angleX = buffer.g2Alt1()
+        val angleY = buffer.g2Alt2()
+        val angleX = buffer.g2Alt3()
         return EventCameraPosition(
             angleX,
             angleY,
