+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/events/EventKeyboardDecoder.kt
@@ -18,7 +18,7 @@ public class EventKeyboardDecoder : ProxyMessageDecoder<EventKeyboard> {
         val keys = ByteArray(count)
         var lastTransmittedKeyPress: Int = -1
         for (i in 0..<count) {
-            val delta = buffer.g3()
+            val delta = buffer.g3Alt2()
             val key = buffer.g1()
             if (i == 0) {
                 lastTransmittedKeyPress = delta
