+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/events/EventNativeMouseClickDecoder.kt
@@ -14,9 +14,9 @@ public class EventNativeMouseClickDecoder : ProxyMessageDecoder<EventNativeMouse
         buffer: JagByteBuf,
         session: Session,
     ): EventNativeMouseClick {
+        val packedCoord = buffer.g4Alt2()
         val lastTransmittedMouseClick = buffer.g2Alt3()
-        val packedCoord = buffer.g4()
-        val code = buffer.g1Alt1()
+        val code = buffer.g1()
         return EventNativeMouseClick(
             lastTransmittedMouseClick,
             code,
