+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/locs/OpLoc3Decoder.kt
@@ -14,10 +14,10 @@ public class OpLoc3Decoder : ProxyMessageDecoder<OpLoc> {
         buffer: JagByteBuf,
         session: Session,
     ): OpLoc {
-        val x = buffer.g2Alt2()
-        val z = buffer.g2()
-        val id = buffer.g2Alt2()
-        val controlKey = buffer.g1() == 1
+        val x = buffer.g2()
+        val controlKey = buffer.g1Alt3() == 1
+        val z = buffer.g2Alt3()
+        val id = buffer.g2()
         return OpLoc(
             id,
             x,
