+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/locs/OpLoc6Decoder.kt
@@ -14,7 +14,7 @@ public class OpLoc6Decoder : ProxyMessageDecoder<OpLoc6> {
         buffer: JagByteBuf,
         session: Session,
     ): OpLoc6 {
-        val id = buffer.g2Alt1()
+        val id = buffer.g2Alt2()
         return OpLoc6(id)
     }
 }
