+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/locs/OpLocTDecoder.kt
@@ -15,13 +15,13 @@ public class OpLocTDecoder : ProxyMessageDecoder<OpLocT> {
         buffer: JagByteBuf,
         session: Session,
     ): OpLocT {
+        val selectedSub = buffer.g2()
+        val selectedObj = buffer.g2Alt1()
         val x = buffer.g2Alt3()
-        val selectedSub = buffer.g2Alt2()
-        val id = buffer.g2Alt2()
-        val selectedObj = buffer.g2Alt2()
-        val controlKey = buffer.g1Alt1() == 1
+        val controlKey = buffer.g1() == 1
+        val z = buffer.g2()
         val selectedCombinedId = buffer.gCombinedIdAlt2()
-        val z = buffer.g2Alt1()
+        val id = buffer.g2Alt2()
         return OpLocT(
             id,
             x,
