+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/misc/user/ClickWorldMapDecoder.kt
@@ -15,7 +15,7 @@ public class ClickWorldMapDecoder : ProxyMessageDecoder<ClickWorldMap> {
         buffer: JagByteBuf,
         session: Session,
     ): ClickWorldMap {
-        val packed = buffer.g4Alt2()
+        val packed = buffer.g4Alt3()
         return ClickWorldMap(CoordGrid(packed))
     }
 }
