+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/misc/user/MoveGameClickDecoder.kt
@@ -14,9 +14,9 @@ public class MoveGameClickDecoder : ProxyMessageDecoder<MoveGameClick> {
         buffer: JagByteBuf,
         session: Session,
     ): MoveGameClick {
+        val x = buffer.g2Alt3()
         val keyCombination = buffer.g1Alt3()
-        val z = buffer.g2()
-        val x = buffer.g2Alt1()
+        val z = buffer.g2Alt3()
         return MoveGameClick(
             x,
             z,
