+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/misc/user/MoveMinimapClickDecoder.kt
@@ -15,9 +15,9 @@ public class MoveMinimapClickDecoder : ProxyMessageDecoder<MoveMinimapClick> {
         session: Session,
     ): MoveMinimapClick {
         // The x, z and keyCombination get scrambled between revisions
+        val x = buffer.g2Alt3()
         val keyCombination = buffer.g1Alt3()
-        val z = buffer.g2()
-        val x = buffer.g2Alt1()
+        val z = buffer.g2Alt3()
 
         // The arguments below are consistent across revisions
         val minimapWidth = buffer.g1()
