+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/misc/user/SetHeadingDecoder.kt
@@ -14,7 +14,7 @@ public class SetHeadingDecoder : ProxyMessageDecoder<SetHeading> {
         buffer: JagByteBuf,
         session: Session,
     ): SetHeading {
-        val heading = buffer.g1Alt2()
+        val heading = buffer.g1()
         return SetHeading(heading)
     }
 }
