+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/misc/user/TeleportDecoder.kt
@@ -14,10 +14,10 @@ public class TeleportDecoder : ProxyMessageDecoder<Teleport> {
         buffer: JagByteBuf,
         session: Session,
     ): Teleport {
-        val x = buffer.g2()
+        val z = buffer.g2Alt3()
         val level = buffer.g1Alt1()
-        val z = buffer.g2Alt2()
         val oculusSyncValue = buffer.g4Alt1()
+        val x = buffer.g2Alt1()
         return Teleport(
             oculusSyncValue,
             x,
