+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/npcs/OpNpc2Decoder.kt
@@ -14,8 +14,8 @@ public class OpNpc2Decoder : ProxyMessageDecoder<OpNpc> {
         buffer: JagByteBuf,
         session: Session,
     ): OpNpc {
-        val index = buffer.g2Alt3()
-        val controlKey = buffer.g1() == 1
+        val controlKey = buffer.g1Alt3() == 1
+        val index = buffer.g2()
         return OpNpc(
             index,
             controlKey,
