+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/objs/OpObj1Decoder.kt
@@ -15,9 +15,9 @@ public class OpObj1Decoder : ProxyMessageDecoder<OpObj> {
         session: Session,
     ): OpObj {
         val id = buffer.g2Alt2()
-        val controlKey = buffer.g1Alt2() == 1
-        val z = buffer.g2()
-        val x = buffer.g2Alt2()
+        val z = buffer.g2Alt1()
+        val x = buffer.g2Alt1()
+        val controlKey = buffer.g1Alt1() == 1
         return OpObj(
             id,
             x,
