+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/objs/OpObj2Decoder.kt
@@ -14,10 +14,10 @@ public class OpObj2Decoder : ProxyMessageDecoder<OpObj> {
         buffer: JagByteBuf,
         session: Session,
     ): OpObj {
-        val x = buffer.g2Alt3()
-        val id = buffer.g2Alt3()
-        val controlKey = buffer.g1Alt2() == 1
-        val z = buffer.g2Alt3()
+        val controlKey = buffer.g1Alt3() == 1
+        val z = buffer.g2Alt2()
+        val x = buffer.g2Alt2()
+        val id = buffer.g2()
         return OpObj(
             id,
             x,
