+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/objs/OpObj3Decoder.kt
@@ -14,10 +14,10 @@ public class OpObj3Decoder : ProxyMessageDecoder<OpObj> {
         buffer: JagByteBuf,
         session: Session,
     ): OpObj {
-        val controlKey = buffer.g1Alt1() == 1
+        val controlKey = buffer.g1Alt2() == 1
+        val id = buffer.g2Alt2()
         val x = buffer.g2Alt2()
-        val z = buffer.g2Alt3()
-        val id = buffer.g2Alt1()
+        val z = buffer.g2Alt1()
         return OpObj(
             id,
             x,
