+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/objs/OpObj4Decoder.kt
@@ -14,10 +14,10 @@ public class OpObj4Decoder : ProxyMessageDecoder<OpObj> {
         buffer: JagByteBuf,
         session: Session,
     ): OpObj {
-        val x = buffer.g2Alt2()
-        val controlKey = buffer.g1() == 1
-        val id = buffer.g2Alt2()
-        val z = buffer.g2Alt1()
+        val x = buffer.g2Alt3()
+        val controlKey = buffer.g1Alt3() == 1
+        val z = buffer.g2Alt2()
+        val id = buffer.g2Alt3()
         return OpObj(
             id,
             x,
