+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/objs/OpObj5Decoder.kt
@@ -14,10 +14,10 @@ public class OpObj5Decoder : ProxyMessageDecoder<OpObj> {
         buffer: JagByteBuf,
         session: Session,
     ): OpObj {
-        val z = buffer.g2()
-        val controlKey = buffer.g1Alt2() == 1
-        val x = buffer.g2Alt1()
         val id = buffer.g2Alt1()
+        val z = buffer.g2()
+        val x = buffer.g2()
+        val controlKey = buffer.g1Alt3() == 1
         return OpObj(
             id,
             x,
