+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/objs/OpObj6Decoder.kt
@@ -14,9 +14,9 @@ public class OpObj6Decoder : ProxyMessageDecoder<OpObj6> {
         buffer: JagByteBuf,
         session: Session,
     ): OpObj6 {
+        val id = buffer.g2Alt2()
         val z = buffer.g2Alt2()
-        val id = buffer.g2()
-        val x = buffer.g2Alt1()
+        val x = buffer.g2()
         return OpObj6(
             id,
             x,
