+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/players/OpPlayer4Decoder.kt
@@ -15,7 +15,7 @@ public class OpPlayer4Decoder : ProxyMessageDecoder<OpPlayer> {
         session: Session,
     ): OpPlayer {
         val index = buffer.g2Alt3()
-        val controlKey = buffer.g1Alt2() == 1
+        val controlKey = buffer.g1Alt1() == 1
         return OpPlayer(
             index,
             controlKey,
