+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/players/OpPlayer6Decoder.kt
@@ -15,7 +15,7 @@ public class OpPlayer6Decoder : ProxyMessageDecoder<OpPlayer> {
         session: Session,
     ): OpPlayer {
         val index = buffer.g2()
-        val controlKey = buffer.g1Alt3() == 1
+        val controlKey = buffer.g1Alt1() == 1
         return OpPlayer(
             index,
             controlKey,
