+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/players/OpPlayer8Decoder.kt
@@ -14,8 +14,8 @@ public class OpPlayer8Decoder : ProxyMessageDecoder<OpPlayer> {
         buffer: JagByteBuf,
         session: Session,
     ): OpPlayer {
-        val controlKey = buffer.g1Alt1() == 1
         val index = buffer.g2Alt3()
+        val controlKey = buffer.g1Alt1() == 1
         return OpPlayer(
             index,
             controlKey,
