+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/players/OpPlayerTDecoder.kt
@@ -2,7 +2,7 @@ package net.rsprox.protocol.v230.game.incoming.decoder.codec.players
 
 import net.rsprot.buffer.JagByteBuf
 import net.rsprot.protocol.ClientProt
-import net.rsprot.protocol.util.gCombinedIdAlt1
+import net.rsprot.protocol.util.gCombinedIdAlt3
 import net.rsprox.protocol.ProxyMessageDecoder
 import net.rsprox.protocol.game.incoming.model.players.OpPlayerT
 import net.rsprox.protocol.session.Session
@@ -15,11 +15,11 @@ public class OpPlayerTDecoder : ProxyMessageDecoder<OpPlayerT> {
         buffer: JagByteBuf,
         session: Session,
     ): OpPlayerT {
-        val index = buffer.g2Alt3()
-        val selectedCombinedId = buffer.gCombinedIdAlt1()
-        val selectedObj = buffer.g2Alt2()
-        val selectedSub = buffer.g2()
-        val controlKey = buffer.g1() == 1
+        val selectedCombinedId = buffer.gCombinedIdAlt3()
+        val selectedSub = buffer.g2Alt1()
+        val selectedObj = buffer.g2()
+        val index = buffer.g2Alt2()
+        val controlKey = buffer.g1Alt2() == 1
         return OpPlayerT(
             index,
             controlKey,
