+++ protocol/osrs-230/src/main/kotlin/net/rsprox/protocol/v230/game/incoming/decoder/codec/resumed/ResumePauseButtonDecoder.kt
@@ -2,7 +2,7 @@ package net.rsprox.protocol.v230.game.incoming.decoder.codec.resumed
 
 import net.rsprot.buffer.JagByteBuf
 import net.rsprot.protocol.ClientProt
-import net.rsprot.protocol.util.gCombinedIdAlt1
+import net.rsprot.protocol.util.gCombinedId
 import net.rsprox.protocol.ProxyMessageDecoder
 import net.rsprox.protocol.game.incoming.model.resumed.ResumePauseButton
 import net.rsprox.protocol.session.Session
@@ -16,7 +16,7 @@ public class ResumePauseButtonDecoder : ProxyMessageDecoder<ResumePauseButton> {
         session: Session,
     ): ResumePauseButton {
         val sub = buffer.g2Alt2()
-        val combinedId = buffer.gCombinedIdAlt1()
+        val combinedId = buffer.gCombinedId()
         return ResumePauseButton(
             combinedId,
             sub,
